+++ src/test/java/dev/lajoscseppento/gradle/plugin/common/property/BooleanSystemPropertyTest.java
@@ -10,11 +10,13 @@ import org.junitpioneer.jupiter.SetSystemProperty;
 class BooleanSystemPropertyTest {
   private BooleanSystemProperty property1;
   private BooleanSystemProperty property2;
+  private BooleanSystemProperty property3;
 
   @BeforeEach
   void setUp() {
     property1 = new BooleanSystemProperty("property.boolean1", false);
     property2 = new BooleanSystemProperty("property.boolean2", () -> true);
+    property3 = new BooleanSystemProperty("property.boolean3", () -> null);
   }
 
   @Test
@@ -45,6 +47,15 @@ class BooleanSystemPropertyTest {
     assertThatThrownBy(() -> property1.get())
         .isExactlyInstanceOf(InvalidPropertyValueException.class)
         .hasMessage("Cannot parse value of system property property.boolean1 as boolean: abc")
+        .hasNoCause();
+  }
+
+  @Test
+  void testNullDefaultValue() {
+    assertThatThrownBy(() -> property3.get())
+        .isExactlyInstanceOf(InvalidPropertyValueException.class)
+        .hasMessage(
+            "System property property.boolean3 has a null value, but a primitive type was requested (BooleanSystemProperty)")
         .hasNoCause();
   }
 }